% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18nInput.R
\name{i18nInput}
\alias{i18nInput}
\title{Create a i18nInput}
\usage{
i18nInput(inputId, choices)
}
\arguments{
\item{inputId}{A character string.}

\item{choices}{A character vector with language options.}
}
\description{
Creates a widget to use alongside the shiny.i18n package.
}
