% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18nInput.R
\name{updateI18nInput}
\alias{updateI18nInput}
\title{Update a i18nInput}
\usage{
updateI18nInput(session = getDefaultReactiveDomain(), inputId, selected = NULL)
}
\arguments{
\item{session}{A shiny session object.}

\item{inputId}{A character string.}

\item{selected}{A character string with the selected language.}
}
\description{
Update a i18nInput
}
