% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18nInput.R
\name{updateI18nInput}
\alias{updateI18nInput}
\title{Update i18nInput}
\usage{
updateI18nInput(session = getDefaultReactiveDomain(), inputId, selected = NULL)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{A character string representing the ID of the input element.}

\item{selected}{A character string with the selected language.}
}
\value{
None. This function is used for its side effects.
}
\description{
This function updates the internationalization (i18n) input widget with new language options.
}
